Refactor/w2 : modifer 순서 변경, dataclsss로묶기, 패키지 변경
