+++ app/src/main/java/com/sopt/dive/feature/home/HomeScreen.kt
@@ -1,25 +1,18 @@
 package com.sopt.dive.feature.home
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.sopt.dive.feature.home.components.BirthdayItemComponent
-import com.sopt.dive.feature.home.components.FriendProfileComponent
-import com.sopt.dive.feature.home.components.MyProfileComponent
+import com.sopt.dive.feature.home.components.profile.BirthdayItemComponent
+import com.sopt.dive.feature.home.components.profile.FriendProfileComponent
+import com.sopt.dive.feature.home.components.profile.MyProfileComponent
 
 
 /**
