+++ app/src/main/java/com/sopt/dive/feature/home/components/profile/BirthdayItemComponent.kt
@@ -1,8 +1,7 @@
-package com.sopt.dive.feature.home.components
+package com.sopt.dive.feature.home.components.profile
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
@@ -13,26 +12,21 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
-
 @Composable
-fun MyProfileComponent(userId: String, userNickname: String) {
+fun BirthdayItemComponent(name: String) {
     Box(
         modifier = Modifier
             .fillMaxWidth()
-            .background(Color(0xFFE0F7FA))
+            .background(Color(0xFFFFF9C4))
             .padding(16.dp)
     ) {
-        Column {
-            Text(text = "내 프로필", fontWeight = FontWeight.Bold, color = Color.Black)
-            Text(text = userNickname, color = Color.Black)
-            Text(text = "ID: $userId", color = Color.Gray)
-        }
+        Text(text = "오늘은 $name 님의 생일입니다!", fontWeight = FontWeight.Bold, color = Color.Black)
     }
 }
 
 @Preview (showBackground = true)
 @Composable
-fun MyProfileComponentTest()
+fun BirthdayItemComponentTest()
 {
-    MyProfileComponent(userId = "1234", userNickname = "text")
+    BirthdayItemComponent(name = "SOPT37기")
 }
