+++ app/src/main/java/com/sopt/dive/feature/home/components/profile/MyProfileComponent.kt
@@ -1,7 +1,8 @@
-package com.sopt.dive.feature.home.components
+package com.sopt.dive.feature.home.components.profile
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
@@ -12,21 +13,26 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
+
 @Composable
-fun BirthdayItemComponent(name: String) {
+fun MyProfileComponent(userId: String, userNickname: String) {
     Box(
         modifier = Modifier
             .fillMaxWidth()
-            .background(Color(0xFFFFF9C4))
+            .background(Color(0xFFE0F7FA))
             .padding(16.dp)
     ) {
-        Text(text = "오늘은 $name 님의 생일입니다!", fontWeight = FontWeight.Bold, color = Color.Black)
+        Column {
+            Text(text = "내 프로필", fontWeight = FontWeight.Bold, color = Color.Black)
+            Text(text = userNickname, color = Color.Black)
+            Text(text = "ID: $userId", color = Color.Gray)
+        }
     }
 }
 
 @Preview (showBackground = true)
 @Composable
-fun BirthdayItemComponentTest()
+fun MyProfileComponentTest()
 {
-    BirthdayItemComponent(name = "SOPT37기")
+    MyProfileComponent(userId = "1234", userNickname = "text")
 }
